requirements.txt
python-telegram-bot>=20.0
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pypdf
Pillow
requests
PyMuPDF
beautifulsoup4
google.generative.ai